+++ HISTORY.rst
@@ -156,6 +156,8 @@
 - Consolidate the workflow state action code into a subset of the Workflow
   State model.
 - Store workflow state action errors into the error log of the document.
+- Update the convert workflow action to add transformations to support
+  templating for the argument of the transformation.
 
 4.6.3 (2024-03-28)
 ==================